Split declaration and definition in the single header into separate files
